update vehicle data from int to double

+++ src/app/arbiter.cpp
@@ -201,7 +201,7 @@ void Arbiter::send_openauto_full_screen(bool fullscreen)
     emit openauto_full_screen(fullscreen);
 }
 
-void Arbiter::send_vehicle_data(QString gauge_id, int value)
+void Arbiter::send_vehicle_data(QString gauge_id, double value)
 {
     emit vehicle_update_data(gauge_id, value);
 }
